Rename rendering module to highlighting
