+++ src/highlighting/renderer.rs
