+++ src/highlighting/terminal.rs
