+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod formatting;
+pub mod highlighting;
 pub mod language;
 pub mod parsing;
 pub mod regex;
-pub mod rendering;
 pub mod templating;
