+++ tests/formatting/golden.rs
@@ -2,8 +2,8 @@ use std::fs;
 use std::path::Path;
 
 use technique::formatting::Identity;
+use technique::highlighting::render;
 use technique::parsing;
-use technique::rendering::render;
 
 /// Golden test for the format command
 ///
